requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
chromedriver-py==141.0.7390.37b0
configparser==7.2.0
configreader==0.0.6
openpyxl==3.1.5
pyparsing==2.4.7
pytest==8.3.5
pytest-ast-transformer==1.0.3
pytest-bdd==8.1.0
pytest-check==2.5.3
pytest-cov==6.1.1
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-order==1.3.0
pytest-rerunfailures==15.0
pytest-soft-assertions==0.1.2
pytest-xdist==3.6.1
selenium==4.30.0
webdriver-manager==4.0.2
